packet: read full date field from pkt
The pkt_get_date() treated date field of pkt as a NULL-terminated
string. Usually, this is the case, but actually it does not
correspond to the FTS, according to which the date has a fixed 20
bytes length.

In the field there are cases when the date is different and the
reader returns error. It makes sense to accept the message anyway
even with the broken date, so the patch changes the function to
behave according to the FTS.

Recently was reported, that HPT produces something strange and
causes the error:

ftntoss ERROR: wrong date size in message header (1 bytes instead 20)

The broken pkt hasn't been yet checked. It's possible, that the
function will be adopted again to use the date.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>